feat: refactor helper functions into GovernanceUtils.ts
