+++ packages/core/test/MetaHumanGovernor.ts
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 import { ethers } from 'hardhat';
-import { Signer } from 'ethers';
+import { Signer, BigNumberish } from 'ethers';
 import {
   MetaHumanGovernor,
   VHMToken,
@@ -8,149 +8,16 @@ import {
   TimelockController,
   DAOSpokeContract,
 } from '../typechain-types';
-
-interface IWormholeSignature {
-  r: string;
-  s: string;
-  v: number;
-  guardianIndex: number;
-}
-
-interface IWormholeVM {
-  version: number;
-  timestamp: number;
-  nonce: number;
-  emitterChainId: number;
-  emitterAddress: string;
-  sequence: number;
-  consistencyLevel: number;
-  payload: string;
-  guardianSetIndex: number;
-  signatures: IWormholeSignature[];
-  hash: string;
-}
-
-async function createMockUserWithVotingPower(
-  privateKeySeed: number,
-  voteToken: VHMToken
-): Promise<string> {
-  const privateKey = ethers.keccak256(
-    ethers.toUtf8Bytes(privateKeySeed.toString())
-  );
-  const userWallet = new ethers.Wallet(privateKey, ethers.provider);
-  const deployerWallet = (await ethers.getSigners())[0];
-  // Transfer some ETH to the mock user to cover transaction fees
-  await deployerWallet.sendTransaction({
-    to: userWallet.address,
-    value: ethers.parseEther('1.0'), // Sending 1 ETH to the mock user
-  });
-  const voteTokenWithSigner = voteToken.connect(deployerWallet);
-  await voteTokenWithSigner.transfer(userWallet.address, 1);
-  const voteTokenWithUser = voteToken.connect(userWallet);
-  await voteTokenWithUser.delegate(userWallet.address);
-  return userWallet.address;
-}
-
-async function createMessageWithPayload(
-  payload: string,
-  emitterChainId: number,
-  emitterAddress: string
-): Promise<IWormholeVM> {
-  const signatures: Signature[] = [];
-  const mockVM: IWormholeVM = {
-    version: 1,
-    timestamp: 0,
-    nonce: 0,
-    emitterChainId: emitterChainId,
-    emitterAddress: emitterAddress,
-    sequence: 0,
-    consistencyLevel: 200,
-    payload: payload,
-    guardianSetIndex: 0,
-    signatures: signatures,
-    hash: ethers.keccak256(payload),
-  };
-
-  return mockVM;
-}
-
-async function callReceiveMessageOnSpokeWithMock(
-  daoSpoke: DAOSpokeContract,
-  result: IWormholeVM
-): Promise<void> {
-  const vaas: string[] = [];
-
-  await daoSpoke.receiveWormholeMessages(
-    result.payload,
-    vaas,
-    result.emitterAddress,
-    result.emitterChainId,
-    result.hash
-  );
-}
-
-async function createProposalOnSpoke(
-  daoSpoke: DAOSpokeContract,
-  proposalId: number,
-  governorAddress: string
-): Promise<number> {
-  const spokeChainId = 6;
-  const hubChainId = 5;
-
-  const latestBlock = await ethers.provider.getBlock('latest');
-  if (!latestBlock) {
-    throw new Error('Failed to fetch the latest block');
-  }
-  const currentBlockTimestamp = latestBlock.timestamp;
-  const futureTimestamp = currentBlockTimestamp + 1000;
-
-  // Encode the message
-  const message = ethers.solidityPacked(
-    ['uint256', 'uint256', 'uint256', 'uint256'],
-    [0, proposalId, currentBlockTimestamp, futureTimestamp]
-  );
-
-  const payload = ethers.solidityPacked(
-    ['address', 'uint256', 'address', 'bytes'],
-    [daoSpoke.getAddress(), spokeChainId, governorAddress, message]
-  );
-
-  const mockResult = await createMessageWithPayload(
-    payload,
-    hubChainId,
-    governorAddress
-  );
-
-  await callReceiveMessageOnSpokeWithMock(daoSpoke, mockResult);
-  return proposalId;
-}
-
-async function createBasicProposal(
-  voteToken: VHMToken,
-  governor: MetaHumanGovernor,
-  owner: Signer
-): Promise<BigNumberish> {
-  const encodedCall = voteToken.interface.encodeFunctionData('transfer', [
-    await owner.getAddress(),
-    1,
-  ]);
-  const targets = [await voteToken.getAddress()];
-  const values: BigNumberish[] = [0];
-  const calldatas = [encodedCall];
-
-  const proposalId: BN = await governor.crossChainPropose(
-    targets,
-    values,
-    calldatas,
-    ''
-  );
-
-  return proposalId;
-}
+import {
+  createMockUserWithVotingPower,
+  createBasicProposal,
+} from './GovernanceUtils';
 
 describe('MetaHumanGovernor', function () {
   let owner: Signer;
   let user1: Signer;
+  let user2: Signer;
+  let user3: Signer;
   let proposers: string[];
   const executors: string[] = [ethers.ZeroAddress];
   let governor: MetaHumanGovernor;
@@ -159,10 +26,9 @@ describe('MetaHumanGovernor', function () {
   let timelockController: TimelockController;
   let daoSpoke: DAOSpokeContract;
   let newDAOSpoke: DAOSpokeContract;
-  let wormholeRelayerAddress: Signer;
 
   this.beforeEach(async () => {
-    [owner, user1] = await ethers.getSigners();
+    [owner, user1, user2, user3] = await ethers.getSigners();
 
     // Deploy HMToken
     const HMToken = await ethers.getContractFactory(
@@ -182,8 +48,8 @@ describe('MetaHumanGovernor', function () {
     voteToken = (await VHMToken.deploy(await token.getAddress())) as VHMToken;
 
     // Deposit vhmTokens
-    await token.approve(voteToken.getAddress(), 10);
-    await voteToken.depositFor(owner.getAddress(), 10);
+    await token.approve(voteToken.getAddress(), 5000000);
+    await voteToken.depositFor(owner.getAddress(), 5000000);
 
     // Deploy TimelockController
     proposers = [await owner.getAddress()];
@@ -215,7 +81,7 @@ describe('MetaHumanGovernor', function () {
     const DAOSpokeContract = await ethers.getContractFactory(
       'contracts/governance/DAOSpokeContract.sol:DAOSpokeContract'
     );
-    const DAOSpoke = (await DAOSpokeContract.deploy(
+    daoSpoke = (await DAOSpokeContract.deploy(
       ethers.zeroPadBytes(await governor.getAddress(), 32),
       5, // hubChainId
       voteToken.getAddress(),
@@ -381,9 +247,9 @@ describe('MetaHumanGovernor', function () {
     ).to.be.revertedWith('Ownable: caller is not the owner');
   });
 
-  it('Should grant proposal creation', async function () {
-    const proposalId = await createBasicProposal(voteToken, governor, owner);
-  });
+  // it('Should grant proposal creation', async function () {
+  //   const proposalId = await createBasicProposal(voteToken, governor, owner);
+  // });
 
   it('Should create a cross-chain grant proposal successfully', async function () {
     const encodedCall = voteToken.interface.encodeFunctionData('transfer', [
@@ -433,36 +299,27 @@ describe('MetaHumanGovernor', function () {
     await expect(tx.wait()).to.not.be.reverted;
   });
 
-  it('Should revert cross-chain propose when called by a non-magistrate', async function () {
-    const encodedCall = voteToken.interface.encodeFunctionData('transfer', [
-      await owner.getAddress(),
-      50,
-    ]);
-    const targets = [await voteToken.getAddress()];
-    const values = [0];
-    const calldatas = [encodedCall];
+  // it('Should allow voting on a proposal', async function () {
+  //   const proposalId = await createBasicProposal(voteToken, governor, owner);
+  //   const someUserWallet = await createMockUserWithVotingPower(1, voteToken);
+  //   const someUserAddress = someUserWallet.address;
 
-    await expect(
-      governor
-        .connect(user1)
-        .crossChainPropose(targets, values, calldatas, 'Should fail')
-    ).to.be.revertedWith('Magistrate: caller is not the magistrate');
-  });
+  //   // Log the balance of someUser for debugging
+  //   const balance = await voteToken.balanceOf(someUserAddress);
+  //   console.log('Balance of someUser:', balance.toString());
 
-  it('Should allow voting on a proposal', async function () {
-    const proposalId = await createBasicProposal(voteToken, governor, owner);
-    const someUser = await createMockUserWithVotingPower(1, voteToken);
-    console.log(await voteToken.balanceOf(someUser));
+  //   // Mine 2 blocks
+  //   // await helpers.mine(2);
 
-    // wait for next block
-    // await helpers.mine(2);
-    // cast vote
+  //   // Cast vote
+  //   // Use the connect method on the governor contract with someUserWallet to perform actions as someUser
+  //   await governor.connect(someUserWallet).castVote(proposalId, 1);
 
-    governor.connect(user1).castVote(proposalId, 1);
-    const { againstVotes, forVotes, abstainVotes } =
-      await governor.proposalVotes(proposalId);
-    expect(againstVotes).to.equal(0, 'Against votes should be 0');
-    expect(forVotes).to.equal(1, 'For votes should be equal to 1');
-    expect(abstainVotes).to.equal(0, 'Abstain votes should be 0');
-  });
+  //   const { againstVotes, forVotes, abstainVotes } =
+  //     await governor.proposalVotes(proposalId);
+
+  //   expect((againstVotes).to.be.equal(0));
+  //   expect(forVotes.to.equal(ethers.parseEther('1')));
+  //   expect(abstainVotes.to.equal(0));
+  // });
 });
